feat: add and update doc comments, add prepared statement cache limiting, simplify copy in

+++ postgresql/src/main/kotlin/com/github/clasicrando/postgresql/connection/PgConnectOptions.kt
@@ -17,7 +17,7 @@ data class PgConnectOptions(
     val password: String? = null,
     val database: String? = null,
     val logSettings: LogSettings = LogSettings.DEFAULT,
-    val statementCacheCapacity: ULong = 100U,
+    val statementCacheCapacity: UShort = 100U,
     val useExtendedProtocolForSimpleQueries: Boolean = true,
     val extraFloatDigits: String = "2",
     val sslMode: SslMode = SslMode.DEFAULT,
